refactor(lint-reason-from-comment): split into submodules + inherit comment-span ctxt
Two changes triggered by Copilot review feedback on #129:

1. Split the rule's 900-line flat file into the catalogue's standard
   `<rule>.rs` + `<rule>/<concern>.rs` layout:
   - `lint_reason_from_comment.rs` keeps the declaration, config,
     `LintReasonFromComment` state, `register_*`, the `EarlyLintPass`
     driver, and the `cfg_attr_trace` book-keeping.
   - `scan.rs` owns `Comment`, `find_trailing_comment`,
     `find_leading_comment`, and `normalise_comment_text`.
   - `insertion.rs` owns `Insertion`, `build_reason_insertion`,
     `locate_outermost_parens`, and `escape_for_rust_string`.
   - `emit.rs` owns `LintReasonFromComment::{check, emit}` and the
     `file_span` helper.
   Unit tests follow the items they exercise into the submodule
   they now live in.

2. `file_span` previously built spans with `SyntaxContext::root()` /
   `parent = None`, which strips the macro-expansion context of the
   underlying attribute and can defeat
   `report_in_external_macro: false`. Take the context+parent from
   the anchor span (the attribute's own `invocation_span`) instead,
   so comment-derived diagnostic/edit spans participate in the same
   filtering and align with the args-insertion span built from
   `invocation_span.with_lo/with_hi`.
